+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Updated tools:
 
 Misc:
 
+- **HAML-Lint** Add `parallel` option [#1582](https://github.com/sider/runners/pull/1582)
 - **stylelint** Add `stylelint-config-recommended-scss` [#1557](https://github.com/sider/runners/pull/1557)
 
 ## 0.36.1